requirements.txt
Flask==1.1.2
numpy==1.19.5
Pillow==8.1.2
requests==2.25.1
tensorflow==2.4.1
waitress==2.0.0